Added `PusherSwiftWithEncryption` target/framework.
