+++ .travis.yml
@@ -21,11 +21,14 @@ before_install:
 before_deploy:
   - carthage build --no-skip-current
   - carthage archive PusherSwift
+  - carthage archive PusherSwiftWithEncryption
 deploy:
   provider: releases
   api_key:
     secure: cjodA0I5fTEU60M+W1NNCj18EpL7yt3uxLN1aeexN33HfRoo153oi4jV/FWlPtQ4QC5FhFF+Nbc69V2paOSrI4iht+UB2Hb0jYTuHWv8ecYa33GFksl7zD9BCf8D3RZqIagTzJC7Vf0FqKKlspZMwIMAL02Z5uwsp5h4Zjrq8lo=
-  file: PusherSwift.framework.zip
+  file: 
+    - PusherSwift.framework.zip
+    - PusherSwiftWithEncryption.framework.zip
   skip_cleanup: true
   on:
     repo: pusher/pusher-websocket-swift
@@ -40,5 +43,9 @@ script:
     fi
   - xcodebuild -scheme PusherSwift -destination "$DST" -project PusherSwift.xcodeproj clean build | xcpretty -tc;
   - xcodebuild -scheme PusherSwift -destination "$DST" -project PusherSwift.xcodeproj -enableCodeCoverage YES $ACTION | xcpretty -tc;
+  - if [[ "$DST" != "platform=tvOS"* ]]; then
+      xcodebuild -scheme PusherSwiftWithEncryption -destination "$DST" -project PusherSwift.xcodeproj clean build | xcpretty -tc;
+      xcodebuild -scheme PusherSwiftWithEncryption -destination "$DST" -project PusherSwift.xcodeproj -enableCodeCoverage YES $ACTION | xcpretty -tc;
+    fi
 after_success:
   - sleep 5
