+++ PusherSwift.xcodeproj/xcshareddata/xcschemes/PusherSwiftWithEncryption.xcscheme
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <Scheme
-   LastUpgradeVersion = "0930"
+   LastUpgradeVersion = "1140"
    version = "1.3">
    <BuildAction
       parallelizeBuildables = "YES"
@@ -14,9 +14,9 @@
             buildForAnalyzing = "YES">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "33831C881A9CF61600B124F1"
-               BuildableName = "PusherSwift.framework"
-               BlueprintName = "PusherSwift"
+               BlueprintIdentifier = "73D8A1E42435E5F3001FDE05"
+               BuildableName = "PusherSwiftWithEncryption.framework"
+               BlueprintName = "PusherSwiftWithEncryption"
                ReferencedContainer = "container:PusherSwift.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -26,25 +26,15 @@
       buildConfiguration = "Debug"
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
-      shouldUseLaunchSchemeArgsEnv = "YES"
-      codeCoverageEnabled = "YES">
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "33831C881A9CF61600B124F1"
-            BuildableName = "PusherSwift.framework"
-            BlueprintName = "PusherSwift"
-            ReferencedContainer = "container:PusherSwift.xcodeproj">
-         </BuildableReference>
-      </MacroExpansion>
+      shouldUseLaunchSchemeArgsEnv = "YES">
       <Testables>
          <TestableReference
             skipped = "NO">
             <BuildableReference
                BuildableIdentifier = "primary"
-               BlueprintIdentifier = "33831C931A9CF61600B124F1"
-               BuildableName = "PusherSwiftTests.xctest"
-               BlueprintName = "PusherSwiftTests"
+               BlueprintIdentifier = "73D8A20A2435F24E001FDE05"
+               BuildableName = "PusherSwiftWithEncryptionTests.xctest"
+               BlueprintName = "PusherSwiftWithEncryptionTests"
                ReferencedContainer = "container:PusherSwift.xcodeproj">
             </BuildableReference>
          </TestableReference>
@@ -60,15 +50,6 @@
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES">
-      <MacroExpansion>
-         <BuildableReference
-            BuildableIdentifier = "primary"
-            BlueprintIdentifier = "33831C881A9CF61600B124F1"
-            BuildableName = "PusherSwift.framework"
-            BlueprintName = "PusherSwift"
-            ReferencedContainer = "container:PusherSwift.xcodeproj">
-         </BuildableReference>
-      </MacroExpansion>
    </LaunchAction>
    <ProfileAction
       buildConfiguration = "Release"
@@ -79,9 +60,9 @@
       <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
-            BlueprintIdentifier = "33831C881A9CF61600B124F1"
-            BuildableName = "PusherSwift.framework"
-            BlueprintName = "PusherSwift"
+            BlueprintIdentifier = "73D8A1E42435E5F3001FDE05"
+            BuildableName = "PusherSwiftWithEncryption.framework"
+            BlueprintName = "PusherSwiftWithEncryption"
             ReferencedContainer = "container:PusherSwift.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
