+++ PusherSwiftWithEncryption.podspec
@@ -1,7 +1,7 @@
 Pod::Spec.new do |s|
-  s.name             = 'PusherSwift'
+  s.name             = 'PusherSwiftWithEncryption'
   s.version          = '7.2.0'
-  s.summary          = 'A Pusher client library in Swift'
+  s.summary          = 'A Pusher client library in Swift that supports encrypted channels'
   s.homepage         = 'https://github.com/pusher/pusher-websocket-swift'
   s.license          = 'MIT'
   s.author           = { "Pusher Limited" => "support@pusher.com" }
@@ -10,12 +10,13 @@ Pod::Spec.new do |s|
 
   s.swift_version = '4.2'
   s.requires_arc  = true
-  s.source_files  = ['Sources/*.swift', 'Sources/PusherSwift-Only/*.swift']
+  s.source_files  = ['Sources/*.swift', 'Sources/PusherSwiftWithEncryption-Only/*.swift']
 
   s.dependency 'ReachabilitySwift', '4.3.0'
   s.dependency 'Starscream', '~> 3.0.5'
+  s.dependency 'Sodium', '~> 0.8.0'
 
   s.ios.deployment_target = '8.0'
-  s.osx.deployment_target = '10.10'
+  s.osx.deployment_target = '10.11'
   s.tvos.deployment_target = '9.0'
 end
