+++ Sources/PusherSwift-Only/PusherDecryptor.swift
@@ -1,0 +1,8 @@
+
+class PusherDecryptor {
+    
+    static func decrypt(data dataAsString: String?, keyProvider: PusherKeyProviding?) -> String? {
+        return dataAsString
+    }
+    
+}
