+++ Tests/PresenceChannelTests.swift
@@ -1,5 +1,10 @@
-@testable import PusherSwift
 import XCTest
+
+#if WITH_ENCRYPTION
+    @testable import PusherSwiftWithEncryption
+#else
+    @testable import PusherSwift
+#endif
 
 class PusherPresenceChannelTests: XCTestCase {
     var pusher: Pusher!
