+++ Tests/PusherCryptoTest.swift
@@ -1,6 +1,10 @@
-@testable
-import PusherSwift
 import XCTest
+
+#if WITH_ENCRYPTION
+    @testable import PusherSwiftWithEncryption
+#else
+    @testable import PusherSwift
+#endif
 
 class PusherCryptoTest: XCTestCase {
     func testHMACGeneratorGeneratesCorrectMAC() {
